'Address range already used by ram allocator' bug corrected. It was due to buggy memcmp function. It does not compare the final 4 bites of memory pages

+++ include/pe.hpp
@@ -202,7 +202,7 @@ public:
     static void set_mmio(mword, Paddr);
     
     static void add_pe_state(mword, Paddr, Paddr, Paddr, mword);
-    static void add_pe_state(size_t, int, mword, Paddr, Paddr, Paddr, mword, mword pti);
+    static void add_pe_state(size_t, size_t, mword, Paddr, Paddr, Paddr, mword, mword pti);
     static void add_pe_state(mword, uint8, mword);
     static void set_rip1(mword);
